+++ custom_components/maintenance_supporter/frontend/strategy/chunks/chunk-WETLRHNO.js
@@ -1,4 +1,4 @@
-import{a as r}from"./chunk-XJY2KOM6.js";var a=r`
+import{a as r}from"./chunk-7FWIJBSZ.js";var a=r`
   ha-card { overflow: hidden; }
   .card-content {
     padding: 16px;